+++ src/core.ts
@@ -57,11 +57,13 @@ import { SkyRadioModule } from './modules/radio';
 import { SkyRepeaterModule } from './modules/repeater';
 import { SkySearchModule } from './modules/search';
 import { SkySectionedFormModule } from './modules/sectioned-form';
+import { SkySelectFieldModule } from './modules/select-field';
 import { SkySortModule } from './modules/sort';
 import { SkyTabsModule } from './modules/tabs';
 import { SkyTextExpandModule } from './modules/text-expand';
 import { SkyTextExpandRepeaterModule } from './modules/text-expand-repeater';
 import { SkyTextHighlightModule } from './modules/text-highlight';
+import { SkyToastModule } from './modules/toast';
 import { SkyTokensModule } from './modules/tokens';
 import { SkyToolbarModule } from './modules/toolbar';
 import { SkyTilesModule } from './modules/tiles';
@@ -116,6 +118,7 @@ import { SkyWaitModule } from './modules/wait';
     SkyRepeaterModule,
     SkySearchModule,
     SkySectionedFormModule,
+    SkySelectFieldModule,
     SkySortModule,
     SkyTabsModule,
     SkyTextExpandModule,
@@ -123,6 +126,7 @@ import { SkyWaitModule } from './modules/wait';
     SkyTextHighlightModule,
     SkyTilesModule,
     SkyTimepickerModule,
+    SkyToastModule,
     SkyTokensModule,
     SkyToolbarModule,
     SkyUrlValidationModule,
@@ -179,6 +183,8 @@ export * from './modules/radio';
 export * from './modules/repeater';
 export * from './modules/search';
 export * from './modules/sectioned-form';
+export * from './modules/select-field';
+export * from './modules/select-field/types';
 export * from './modules/sort';
 export * from './modules/tabs';
 export * from './modules/text-expand';
@@ -186,6 +192,7 @@ export * from './modules/text-expand-repeater';
 export * from './modules/text-highlight';
 export * from './modules/tiles';
 export * from './modules/timepicker';
+export * from './modules/toast';
 export * from './modules/tokens';
 export * from './modules/toolbar';
 export * from './modules/url-validation';