+++ src/demos/list-filters/list-filters-inline-demo.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
+import 'rxjs/add/observable/of';
 
 import {
   ListItemModel