+++ src/modules/list-toolbar/list-toolbar.module.ts
@@ -9,12 +9,18 @@ import { SkyListToolbarComponent } from './list-toolbar.component';
 import { SkyListToolbarItemComponent } from './list-toolbar-item.component';
 import { SkyListToolbarItemRendererComponent } from './list-toolbar-item-renderer.component';
 import { SkyListToolbarSortComponent } from './list-toolbar-sort.component';
+
+import {
+  SkyListToolbarViewActionsComponent
+} from './list-toolbar-view-actions.component';
+
 @NgModule({
   declarations: [
     SkyListToolbarComponent,
     SkyListToolbarItemComponent,
     SkyListToolbarItemRendererComponent,
-    SkyListToolbarSortComponent
+    SkyListToolbarSortComponent,
+    SkyListToolbarViewActionsComponent
   ],
   imports: [
     CommonModule,
@@ -28,7 +34,8 @@ import { SkyListToolbarSortComponent } from './list-toolbar-sort.component';
     SkyListToolbarComponent,
     SkyListToolbarItemComponent,
     SkyListToolbarItemRendererComponent,
-    SkyListToolbarSortComponent
+    SkyListToolbarSortComponent,
+    SkyListToolbarViewActionsComponent
   ],
   providers: [
   ]