+++ src/modules/list-view-grid/list-view-grid.component.ts
@@ -12,6 +12,8 @@ import {
 
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
+import 'rxjs/add/operator/distinctUntilChanged';
+import 'rxjs/add/operator/take';
 import 'rxjs/add/operator/takeUntil';
 
 import {