+++ src/modules/toast/toast-body-context.ts
@@ -1,0 +1,10 @@
+// #region imports
+import {
+  Injectable
+} from '@angular/core';
+// #endregion
+
+@Injectable()
+export class SkyToastBodyContext {
+  public message: string;
+}